+++ addons/sale_project/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import product
+from . import project
+from . import sale_order
