+++ addons/sale_project/views/project_task_views.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="project_project_view_tree_inherit_sale_project" model="ir.ui.view">
+        <field name="name">project.project.tree.inherit.sale.project</field>
+        <field name="model">project.project</field>
+        <field name="inherit_id" ref="project.view_project"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='analytic_account_id']" position="after">
+                <field name="sale_order_id" optional="hide" widget="sale_order_many2one"/>
+                <field name="sale_line_id" optional="hide" widget="sale_order_many2one"/>
+            </xpath>
+        </field>
+    </record>
+
+    <record id="view_sale_project_inherit_form" model="ir.ui.view">
+        <field name="name">project.task.view.inherit</field>
+        <field name="model">project.task</field>
+        <field name="groups_id" eval="[(4, ref('base.group_user'))]"/>
+        <field name="inherit_id" ref="project.view_task_form2"/>
+        <field name="arch" type="xml">
+            <div name="button_box" position="inside">
+                <button class="d-none d-md-inline oe_stat_button"
+                        type="object" name="action_view_so" icon="fa-dollar"
+                        attrs="{'invisible': [('sale_order_id', '=', False)]}"
+                        string="Sales Order"
+                        groups="sales_team.group_sale_salesman"/>
+            </div>
+            <xpath expr="//group[@name='sales_order_group']" position="inside">
+                <field name="sale_line_id" string="Sales Order Item" attrs="{'invisible': ['|', ('partner_id', '=', False), '&amp;', ('sale_order_id', '!=', False), ('sale_line_id', '=', False)]}" options='{"no_open": True}' readonly="1" context="{'create': False, 'edit': False, 'delete': False}" groups="sales_team.group_sale_salesman"/>
+                <field name="sale_order_id" attrs="{'invisible': [('project_id', '!=', False)]}" groups="sales_team.group_sale_salesman"/>
+                <field name="commercial_partner_id" invisible="1" />
+                <field name="project_sale_order_id" invisible="1"/>
+            </xpath>
+        </field>
+    </record>
+
+    <record id="project_task_view_form_inherit_sale_line_editable" model="ir.ui.view">
+        <field name="name">project.task.form.inherit.sale.line.editable.salesman</field>
+        <field name="model">project.task</field>
+        <field name="inherit_id" ref="view_sale_project_inherit_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='sale_line_id']" position="attributes">
+                <attribute name="options">{"no_create": True}</attribute>
+                <attribute name="readonly">0</attribute>
+            </xpath>
+        </field>
+        <field name="groups_id" eval="[(4, ref('sales_team.group_sale_salesman'))]"/>
+    </record>
+
+    <record id="project_task_view_search" model="ir.ui.view">
+        <field name="name">project.task.search.inherit</field>
+        <field name="model">project.task</field>
+        <field name="inherit_id" ref="project.view_task_search_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='stage_id']" position="before">
+                <field name="sale_order_id" string="Sale Order" filter_domain="['|', ('sale_order_id', 'ilike', self), ('sale_line_id', 'ilike', self)]"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
