+++ addons/sale_timesheet/security/sale_timesheet_security.xml
@@ -3,17 +3,6 @@
 
     <data noupdate="1">
 
-        <record id="sale_order_line_rule_project_manager" model="ir.rule">
-            <field name="name">Project Manager Sales Orders Line</field>
-            <field name="model_id" ref="sale.model_sale_order_line"/>
-            <field name="domain_force">['&amp;', '&amp;', ('state', 'in', ['sale', 'done']), ('is_service', '=', True), '|', ('project_id','!=', False), ('task_id','!=', False)]</field>
-            <field name="groups" eval="[(4, ref('project.group_project_manager'))]"/>
-            <field name="perm_create" eval="0"/>
-            <field name="perm_write" eval="0"/>
-            <field name="perm_unlink" eval="0"/>
-            <field name="perm_read" eval="1"/>
-        </record>
-
         <!-- Override this rule because in hr_timesheet,
             the lowest access right can only see own timesheets (model: account.analytic.line)
             and this ir.rule accept all account.analytic.line in its domain.
