+++ addons/sale_timesheet/tests/__init__.py
@@ -7,5 +7,4 @@ from . import test_sale_service
 from . import test_project_billing
 from . import test_reinvoice
 from . import test_reporting
-from . import test_child_tasks
 from . import test_project_overview
