+++ addons/sale_timesheet/views/product_views.xml
@@ -11,9 +11,6 @@
             </xpath>
             <xpath expr="//field[@name='service_type']" position="after">
                 <field name="service_policy" widget="radio" attrs="{'invisible': [('type','!=','service')]}"/>
-                <field name="service_tracking" widget="radio" attrs="{'invisible': [('type','!=','service')]}"/>
-                <field name="project_id" attrs="{'invisible':[('service_tracking','!=','task_global_project')], 'required':[('service_tracking','==','task_global_project')]}"/>
-                <field name="project_template_id" context="{'active_test': False}" attrs="{'invisible':[('service_tracking','not in',['task_in_project', 'project_only'])]}"/>
             </xpath>
         </field>
     </record>
