+++ addons/sale_timesheet/views/sale_order_views.xml
@@ -6,26 +6,10 @@
             <field name="inherit_id" ref="sale.view_order_form"/>
             <field name="arch" type="xml">
                 <data>
-                    <xpath expr="//button[@name='action_view_invoice']" position="before">
-                       <button type="object"
-                           name="action_view_project_ids"
-                           class="oe_stat_button"
-                           icon="fa-puzzle-piece"
-                           string="Project Overview"
-                           attrs="{'invisible': [('project_ids', '=', [])]}"
-                           groups="project.group_project_manager">
-                          <field name="project_ids" invisible="1"/>
-                       </button>
+                    <xpath expr="//button[@name='action_view_project_ids']" position="attributes">
+                        <attribute name="string">Project Overview</attribute>
                     </xpath>
                     <xpath expr="//button[@name='action_view_invoice']" position="before">
-                        <button type="object"
-                           name="action_view_task"
-                           class="oe_stat_button"
-                           icon="fa-tasks"
-                           attrs="{'invisible': [('tasks_count', '=', 0)]}"
-                           groups="project.group_project_user">
-                           <field name="tasks_count" widget="statinfo" string="Tasks"/>
-                        </button>
                         <field name="timesheet_count" invisible="1" />
                         <button type="object"
                            name="action_view_timesheet"
@@ -41,10 +25,6 @@
                                 <span class="o_stat_text">Recorded</span>
                             </div>
                         </button>
-                    </xpath>
-                    <xpath expr="//field[@name='analytic_account_id']" position="after">
-                       <field name="visible_project" invisible="1"/>
-                       <field name="project_id" options="{'no_create': True}" attrs="{'invisible': [('visible_project', '=', False)]}"/>
                     </xpath>
                 </data>
            </field>
